Refactored this a lot
I made the iteration for finding operations backwards since it matters for operators like ^ that are not associative. I made it so that expressions enclosed in "()" are considered as a single token. I fixed several syntax errors. It remains to ensure that the function tree is actually correct, because clearly it's not.